+++ module/patch-core.js
@@ -57,13 +57,28 @@ export async function PatchCore() {
   };
   // Patch ActorTokenHelpers.deleteEmbeddedEntity
   const ActorTokenHelpers_deleteEmbeddedEntity = ActorTokenHelpers.prototype.deleteEmbeddedEntity;
-  ActorTokenHelpers.prototype.deleteEmbeddedEntity = async function (embeddedName, id, options = {}) {
-    const item = this.items.get(id);
+  ActorTokenHelpers.prototype.deleteEmbeddedEntity = async function (embeddedName, data, options = {}) {
+    data = data instanceof Array ? data : [data];
+    const ids = data.map((o) => {
+      if (typeof o === "string") return o;
+      return o._id;
+    });
+    const items = this.items.filter((o) => {
+      return ids.includes(o.id);
+    });
 
-    const deleted = await ActorTokenHelpers_deleteEmbeddedEntity.call(this, embeddedName, id, options);
+    // Remove class associations
+    for (let item of items) {
+      if (item.data.type === "class") {
+        await item._onLevelChange(getProperty(item.data, "data.level"), 0);
+      }
+    }
+
+    // Delete item
+    const deleted = await ActorTokenHelpers_deleteEmbeddedEntity.call(this, embeddedName, data, options);
 
     // Remove token effects for deleted buff
-    if (item) {
+    for (let item of items) {
       let promises = [];
       if (item.type === "buff" && item.data.data.active) {
         const isLinkedToken = getProperty(this.data, "token.actorLink");
@@ -80,10 +95,22 @@ export async function PatchCore() {
 
   const ActorTokenHelpers_createEmbeddedEntity = ActorTokenHelpers.prototype.createEmbeddedEntity;
   ActorTokenHelpers.prototype.createEmbeddedEntity = async function (embeddedName, data, options = {}) {
-    const created = await ActorTokenHelpers_createEmbeddedEntity.call(this, embeddedName, data, options);
+    let created = await ActorTokenHelpers_createEmbeddedEntity.call(this, embeddedName, data, options);
+    created = created instanceof Array ? created : [created];
+
     if (embeddedName === "OwnedItem") {
-      if (data.type === "buff" && getProperty(data, "data.active") === true) {
-        this.toggleConditionStatusIcons();
+      for (let i of created) {
+        const item = this.items.get(i._id);
+
+        // Show new buff icon
+        if (item.data.type === "buff" && getProperty(item.data, "data.active") === true) {
+          await this.toggleConditionStatusIcons();
+        }
+
+        // Add class features
+        if (item.data.type === "class") {
+          await item._onLevelChange(0, item.data.data.level);
+        }
       }
     }
 
@@ -92,10 +119,34 @@ export async function PatchCore() {
 
   const ActorTokenHelpers_updateEmbeddedEntity = ActorTokenHelpers.prototype.updateEmbeddedEntity;
   ActorTokenHelpers.prototype.updateEmbeddedEntity = async function (embeddedName, data, options = {}) {
-    const updates = await ActorTokenHelpers_updateEmbeddedEntity.call(this, embeddedName, data, options);
+    data = data instanceof Array ? data : [data];
+    const prevData = data.reduce((cur, o) => {
+      const obj = duplicate(this.items.get(o._id));
+      cur.set(o._id, obj);
+      return cur;
+    }, new Map());
+    let updates = await ActorTokenHelpers_updateEmbeddedEntity.call(this, embeddedName, data, options);
+    updates = updates instanceof Array ? updates : [updates];
+
     if (embeddedName === "OwnedItem") {
-      if (updates.type === "buff" && data["data.active"] !== undefined) {
-        this.toggleConditionStatusIcons();
+      for (let i of updates) {
+        const item = this.items.get(i._id);
+        const prev = prevData.get(i._id);
+        const d = data.find((o) => o._id === i._id);
+
+        // Show updated buff icon
+        if (item.data.type === "buff" && d["data.active"] !== undefined) {
+          await this.toggleConditionStatusIcons();
+        }
+
+        // Add or remove class features
+        if (item.data.type === "class") {
+          const prevLevel = getProperty(prev, "data.level");
+          const newLevel = getProperty(item.data, "data.level");
+          if (prevLevel !== newLevel) {
+            await item._onLevelChange(prevLevel, newLevel);
+          }
+        }
       }
     }